requirements.txt
Flask>=2.3.0,<3.0.0
mysql-connector-python>=8.0.33,<9.0.0
argon2-cffi>=23.1.0,<24.0.0
Werkzeug>=2.3.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
Flask-WTF>=1.1.0,<2.0.0
Flask-Limiter>=3.5.0,<4.0.0
Flask-CORS>=4.0.0,<5.0.0
PyJWT>=2.8.0,<3.0.0
flask-swagger-ui>=4.11.1,<5.0.0
apispec>=6.3.0,<7.0.0
>=3.20.0,<4.0.0

# 开发依赖（可选，使用 pip install -r requirements-dev.txt）
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
